# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TruckLink API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /trucks:
    get:
      operationId: listTrucks
      tags: [trucks]
      summary: List all trucks
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [available, in_transit, maintenance, inactive]
        - name: ownerId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of trucks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Truck"
    post:
      operationId: createTruck
      tags: [trucks]
      summary: Create a truck
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTruckBody"
      responses:
        "201":
          description: Created truck
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Truck"
  /trucks/{id}:
    get:
      operationId: getTruck
      tags: [trucks]
      summary: Get truck by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Truck
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Truck"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateTruck
      tags: [trucks]
      summary: Update truck
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTruckBody"
      responses:
        "200":
          description: Updated truck
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Truck"
    delete:
      operationId: deleteTruck
      tags: [trucks]
      summary: Delete truck
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /trucks/telematics:
    get:
      operationId: getTruckTelematics
      tags: [trucks]
      summary: Get real-time telematics for all trucks
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Truck telematics data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TruckTelematics"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, dispatcher, driver, carrier, shipper]
        company:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        rating:
          type: ["number", "null"]
        totalLoads:
          type: integer
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - name
        - role
        - totalLoads
        - isActive
        - createdAt
    UpdateUserBody:
      type: object
      properties:
        name:
          type: string
        company:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
    Truck:
      type: object
      properties:
        id:
          type: integer
        ownerId:
          type: integer
        make:
          type: string
        model:
          type: string
        year:
          type: integer
        licensePlate:
          type: string
        vin:
          type: ["string", "null"]
        truckType:
          type: string
          enum: [flatbed, dry_van, reefer, tanker, step_deck, lowboy, conestoga, auto_carrier]
        status:
          type: string
          enum: [available, in_transit, maintenance, inactive]
        maxWeight:
          type: number
        maxLength:
          type: number
        color:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        currentLat:
          type: ["number", "null"]
        currentLng:
          type: ["number", "null"]
        currentCity:
          type: ["string", "null"]
        currentState:
          type: ["string", "null"]
        mileage:
          type: integer
        fuelLevel:
          type: ["number", "null"]
        lastServiceDate:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        owner:
          $ref: "#/components/schemas/User"
      required:
        - id
        - ownerId
        - make
        - model
        - year
        - licensePlate
        - truckType
        - status
        - maxWeight
        - maxLength
        - mileage
        - createdAt
        - owner
    CreateTruckBody:
      type: object
      properties:
        make:
          type: string
        model:
          type: string
        year:
          type: integer
        licensePlate:
          type: string
        vin:
          type: ["string", "null"]
        truckType:
          type: string
          enum: [flatbed, dry_van, reefer, tanker, step_deck, lowboy, conestoga, auto_carrier]
        maxWeight:
          type: number
        maxLength:
          type: number
        color:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
      required:
        - make
        - model
        - year
        - licensePlate
        - truckType
        - maxWeight
        - maxLength
    UpdateTruckBody:
      type: object
      properties:
        status:
          type: string
          enum: [available, in_transit, maintenance, inactive]
        currentLat:
          type: ["number", "null"]
        currentLng:
          type: ["number", "null"]
        currentCity:
          type: ["string", "null"]
        currentState:
          type: ["string", "null"]
        fuelLevel:
          type: ["number", "null"]
        mileage:
          type: integer
        lastServiceDate:
          type: ["string", "null"]
...
    TruckTelematics:
      type: object
      properties:
        truckId:
          type: integer
        lat:
          type: number
        lng:
          type: number
        speed:
          type: number
        heading:
          type: number
        fuelLevel:
          type: number
        engineTemp:
          type: number
        status:
          type: string
        lastUpdated:
          type: string
          format: date-time
      required:
        - truckId
        - lat
        - lng
        - speed
        - heading
        - fuelLevel
        - engineTemp
        - status
        - lastUpdated